+++ src/main/java/generiek/api/EnrollmentEndpoint.java
@@ -320,12 +320,15 @@ public class EnrollmentEndpoint {
         Map<String, Object> body = new EnrollmentResult(results).transform();
         HttpEntity<Void> requestEntity = new HttpEntity(body, httpHeaders);
 
+        ResponseEntity exchanged = null;
         try {
-            ResponseEntity exchanged = restTemplate.exchange(resultsURI, HttpMethod.POST, requestEntity, Void.class);
+            exchanged = restTemplate.exchange(resultsURI, HttpMethod.POST, requestEntity, Void.class);
             LOG.debug(String.format("Received answer from %s with status %s", resultsURI, exchanged.getStatusCode()));
             return ResponseEntity.ok().body(exchanged.getBody());
         } catch (HttpStatusCodeException e) {
-            return this.errorResponseEntity("Error from the OOAPI results endpoint for enrolment request:" + enrollmentRequest, e);
+            LOG.error(String.format("Error %s from the OOAPI results endpoint for enrolment request: %s. Message: %s", e.getStatusCode(), enrollmentRequest, e.getMessage()));
+            String responseBody = (exchanged!=null) ? exchanged.getBody().toString() : "No content";
+            return ResponseEntity.status(e.getStatusCode()).body(responseBody);
         }
     }